feat: 팔로잉 취소
- 팔로우 관계가 아니라면 404 예외 발생
- 내가 팔로우 한 사람 중 그 사람이 없기 때문에 Not Found

+++ src/main/java/diary/service/FollowService.java
@@ -9,7 +9,9 @@ import diary.responseDto.FollowResponseDto;
 import diary.responseDto.FollowerResponseDto;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -48,6 +50,10 @@ public class FollowService {
 
         // 팔로우 관계인지 확인
         Follow follow = followRepository.findByFollowerAndFollowee(follower.getId(), followee.getId());
+
+        if (follow == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Follow relationship not found");
+        }
 
         followRepository.delete(follow);
     }
